Feature/Routing Dispatching.
* Now the router only process custom event and product action requests.
* Removed support for scheduled task, processing operations and validations from router and dispatcher.
* application dispatcher moved to router module.
