+++ connect/processors_toolkit/router/router.py
@@ -5,47 +5,11 @@
 #
 from __future__ import annotations
 
-from abc import ABC
 from dataclasses import dataclass
-from typing import Dict, List, Type, Union
+from typing import Dict, List, Type
 
-from connect.eaas.core.responses import (
-    CustomEventResponse,
-    ProcessingResponse,
-    ProductActionResponse,
-    ScheduledExecutionResponse,
-    ValidationResponse,
-)
-from connect.processors_toolkit.transactions.contracts import (
-    CustomEventTransaction,
-    ProcessingTransaction,
-    ProductActionTransaction,
-    ScheduledTransaction,
-    ValidationTransaction,
-)
-from connect.processors_toolkit.requests import RequestBuilder
-
-
-class ProcessNotFound(ProcessingTransaction, ABC):
-    def process(self, request: Union[RequestBuilder, dict]) -> ProcessingResponse:
-        """
-        Handles the request in the case no flow controller match.
-
-        :param request:
-        :return: ProcessingResponse
-        """
-        raise NotImplementedError()
-
-
-class ValidationNotFound(ValidationTransaction):
-    def validate(self, request: Union[RequestBuilder, dict]) -> ValidationResponse:
-        """
-        Handles the request in the case no flow controller match.
-
-        :param request:
-        :return: ValidationResponse
-        """
-        raise NotImplementedError()
+from connect.eaas.core.responses import CustomEventResponse, ProductActionResponse
+from connect.processors_toolkit.transactions.contracts import CustomEventTransaction, ProductActionTransaction
 
 
 class CustomEventNotFound(CustomEventTransaction):
@@ -59,7 +23,7 @@ class CustomEventNotFound(CustomEventTransaction):
         return CustomEventResponse.done(http_status=404)
 
 
-class ActionNotFound(ProductActionTransaction):
+class ProductActionNotFound(ProductActionTransaction):
     def handle(self, request: dict) -> ProductActionResponse:
         """
         Handles the request in case no flow controller match.
@@ -70,32 +34,16 @@ class ActionNotFound(ProductActionTransaction):
         return ProductActionResponse.done(http_status=404)
 
 
-class ScheduledNotFound(ScheduledTransaction):
-    def handle(self, request: dict) -> ScheduledExecutionResponse:
-        """
-        Handles the request in the case no flow controller match.
-
-        :param request:
-        :return: ScheduledExecutionResponse
-        """
-        raise NotImplementedError()
-
-
 @dataclass
 class Route:
     scope: str
     process: str
     name: str
 
-    SCOPE_ASSET = 'asset'
-    SCOPE_TIER_CONFIG = 'tier-config'
     SCOPE_PRODUCT = 'product'
 
-    PROCESS_PROCESS = 'process'
-    PROCESS_VALIDATE = 'validate'
     PROCESS_CUSTOM_EVENT = 'custom-event'
     PROCESS_ACTION = 'action'
-    PROCESS_SCHEDULE = 'schedule'
 
     def __post_init__(self):
         if self.scope not in self.scopes():
@@ -108,14 +56,6 @@ class Route:
             raise ValueError('Invalid route name, must not contains spaces.')
 
     @staticmethod
-    def for_process(scope: str, action: str) -> Route:
-        return Route(scope, Route.PROCESS_PROCESS, action)
-
-    @staticmethod
-    def for_validate(scope: str, action: str) -> Route:
-        return Route(scope, Route.PROCESS_VALIDATE, action)
-
-    @staticmethod
     def for_custom_event(action: str) -> Route:
         return Route(Route.SCOPE_PRODUCT, Route.PROCESS_CUSTOM_EVENT, action)
 
@@ -123,24 +63,15 @@ class Route:
     def for_action(action: str) -> Route:
         return Route(Route.SCOPE_PRODUCT, Route.PROCESS_ACTION, action)
 
-    @staticmethod
-    def for_schedule(action: str) -> Route:
-        return Route(Route.SCOPE_PRODUCT, Route.PROCESS_SCHEDULE, action)
-
     def scopes(self) -> List[str]:
         return [
-            self.SCOPE_ASSET,
-            self.SCOPE_TIER_CONFIG,
             self.SCOPE_PRODUCT,
         ]
 
     def processes(self) -> List[str]:
         return [
-            self.PROCESS_PROCESS,
-            self.PROCESS_VALIDATE,
             self.PROCESS_CUSTOM_EVENT,
             self.PROCESS_ACTION,
-            self.PROCESS_SCHEDULE,
         ]
 
     def not_found(self) -> str:
@@ -154,12 +85,8 @@ class Router:
     def __init__(self, routes: Dict[str, Type], not_found: Dict[str, Type]):
         self.__routes: Dict[str, Type] = routes
         self.__not_found: Dict[str, Type] = {
-            'asset.process': ProcessNotFound,
-            'tier-config.process': ProcessNotFound,
-            'asset.validate': ValidationNotFound,
             'product.custom-event': CustomEventNotFound,
-            'product.action': ActionNotFound,
-            'product.schedule': ScheduledNotFound,
+            'product.action': ProductActionNotFound,
         }
 
         self.__not_found.update(not_found)
